requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv
streamlit
requests
httpx
email-validator
python-multipart
bcrypt
pydantic